Moved krylov_solver and friends to solvers library.
This lightens the load on the core library a bit.
