+++ core/CMakeLists.txt
@@ -20,8 +20,6 @@ add_polymec_library(polymec_core polymec.c arch.c allocators.c logging.c timer.c
                     partition_graph.c partition_point_cloud.c
                     text_buffer.c serializer.c string_utils.c 
                     file_utils.c array_utils.c lookup1.c linear_algebra.c 
-                    bd_matrix.c matrix_sparsity.c krylov_solver.c 
-                    petsc_krylov_solver.c hypre_krylov_solver.c
                     ${CMAKE_CURRENT_BINARY_DIR}/special_functions.f90 
                     special_functions.c 
                     exchanger.c lua_types.c lua_core.c
