+++ core/tests/CMakeLists.txt
@@ -39,7 +39,6 @@ add_polymec_core_test(test_hilbert test_hilbert.c)
 add_polymec_core_test(test_adj_graph test_adj_graph.c)
 add_polymec_core_test(test_newton test_newton.c)
 add_polymec_core_test(test_special_functions test_special_functions.c)
-add_polymec_core_test(test_matrix_sparsity test_matrix_sparsity.c ../../geometry/create_uniform_polymesh.c ../../geometry/create_rectilinear_polymesh.c ../../geometry/cubic_lattice.c ../../geometry/polymesh.c)
 add_polymec_core_test(test_point_cloud test_point_cloud.c)
 add_polymec_core_test(test_polynomial test_polynomial.c)
 add_polymec_core_test(test_least_squares test_least_squares.c)
@@ -52,7 +51,6 @@ add_mpi_polymec_core_test(test_migrator test_migrator.c 1 2 3 4)
 add_mpi_polymec_core_test(test_partition_point_cloud test_partition_point_cloud.c ../../geometry/create_point_lattice.c ../../geometry/cubic_lattice.c 1 2 3 4)
 add_mpi_polymec_core_test(test_parallel_sort test_parallel_sort.c 1 2 3 4)
 add_mpi_polymec_core_test(test_repartition_point_cloud test_repartition_point_cloud.c 1 2 3 4)
-add_mpi_polymec_core_test(test_krylov_solver test_krylov_solver.c 1 2)
 
 # Simple driver test with Lua interpreter.
 include(add_driver_test)
