+++ solvers/bd_matrix.c
@@ -5,8 +5,8 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "core/bd_matrix.h"
 #include "core/linear_algebra.h"
+#include "solvers/bd_matrix.h"
 
 struct bd_matrix_t
 {
