+++ solvers/bd_matrix.h
