+++ solvers/matrix_sparsity.c
@@ -5,7 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "core/matrix_sparsity.h"
+#include "solvers/matrix_sparsity.h"
 
 struct matrix_sparsity_t 
 {
