+++ solvers/tests/create_krylov_factories.c
@@ -5,7 +5,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/.
 
-#include "core/krylov_solver.h"
+#include "solvers/krylov_solver.h"
 
 // These functions seamlessly create factories for PETSc- and HYPRE-based
 // Krylov solvers.
